+++ app/index.tsx
@@ -1,8 +1,6 @@
+import { useAuth } from "@clerk/clerk-expo";
 import { Redirect } from "expo-router";
 import React from "react";
-import { Text } from "react-native";
-import { SafeAreaView } from "react-native-safe-area-context";
-import { useAuth } from "@clerk/clerk-expo";
 
 const Home = () => {
   const { isSignedIn } = useAuth();